Move not-verified status box above navigation tabs
