+++ src/client/components/CompanyLocalHeader/index.jsx
@@ -2,13 +2,14 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import pluralize from 'pluralize'
-import GridCol from '@govuk-react/grid-col'
-import Button from '@govuk-react/button'
-import GridRow from '@govuk-react/grid-row'
-import { SPACING, FONT_SIZE, BREAKPOINTS } from '@govuk-react/constants'
-import Details from '@govuk-react/details'
 import Main from '@govuk-react/main'
-
+import GridCol from '@govuk-react/grid-col'
+import GridRow from '@govuk-react/grid-row'
+import Button from '@govuk-react/button'
+import Paragraph from '@govuk-react/paragraph'
+import WarningText from '@govuk-react/warning-text'
+import Details from '@govuk-react/details'
+import { SPACING, FONT_SIZE, BREAKPOINTS } from '@govuk-react/constants'
 import { GREY_3, PURPLE, BLACK } from '../../../client/utils/colours'
 import LocalHeader from '../../../client/components/LocalHeader/LocalHeader'
 import LocalHeaderHeading from '../../../client/components/LocalHeader/LocalHeaderHeading'
@@ -113,6 +114,10 @@ const StyledMainMuted = styled(Main)`
   }
 `
 
+const StyledLink = styled('a')`
+  margin-bottom: 0;
+`
+
 const CompanyLocalHeader = ({
   breadcrumbs,
   flashMessages,
@@ -121,6 +126,8 @@ const CompanyLocalHeader = ({
   returnUrl,
 }) => {
   const queryString = returnUrl ? `${returnUrl}` : `/companies/${company.id}`
+  const showMatchingPrompt =
+    !company.duns_number && !company.pending_dnb_investigation
   return (
     company && (
       <>
@@ -217,6 +224,38 @@ const CompanyLocalHeader = ({
             </StyledDescription>
           )}
         </LocalHeader>
+
+        {showMatchingPrompt && (
+          <StyledMain>
+            <StatusMessage
+              colour={BLACK}
+              id="ga-company-details-matching-prompt"
+            >
+              <WarningText>
+                Business details on this company record have not been verified
+                and could be wrong.
+              </WarningText>
+              <Details summary="Why verify?">
+                <Paragraph>
+                  Using verified business details from a trusted third-party
+                  supplier means we can keep certain information up to date
+                  automatically. This helps reduce duplicate records, provide a
+                  shared view of complex companies and make it more likely we
+                  can link other data sources together.
+                </Paragraph>
+                <Paragraph>
+                  Verification can often be done in just 4 clicks.
+                </Paragraph>
+              </Details>
+              <Button
+                as={StyledLink}
+                href={urls.companies.match.index(company.id)}
+              >
+                Verify business details
+              </Button>
+            </StatusMessage>
+          </StyledMain>
+        )}
 
         {company.archived && (
           <ArchivePanel
